Use helper script for CtrlP file list in homedir
This uses a new helper script for CtrlP file listings (:CtrlP) in
a vcsh-controlled homedir.  When vcsh hasn't been used to enter
a sub-repo (i.e. vcsh hasn't set $GIT_DIR), the full list of
vcsh-controlled files is returned.  Otherwise, the `git ls-files`
contents for that sub-repo are returned.

This change also works around an issue where vcsh wasn't returning the
same results from `vcsh list-tracked` when in a sub-repo vs. not.