Improve Dot_vim path detection
The prior code was incorrect to use the has('nvim') predicate, since
it's possible to start nvim by directly invoking a .vimrc, which is
exactly what vimpager does.  Now we inspect <sfile> and divine our .vim
directory from that, i.e. is it 'init.vim' or '.vimrc'.  If it's
something else, well how in the heck did we even get here?